Move tests for testing stockflow source to test_askenet_source
